requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
click==8.1.3
Django==4.1.3
django-json==0.27
django-nested-admin==4.0.0
exceptiongroup==1.0.0
greenlet==1.1.3
h11==0.14.0
idna==3.4
outcome==1.2.0
Pillow==9.3.0
playwright==1.27.1
pyee==8.1.0
PySocks==1.7.1
python-monkey-business==1.0.0
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
uvicorn==0.19.0
wsproto==1.2.0
